backend/requirements.txt
# /Users/deangladish/CollabNext_public/backend/requirements.txt

Flask==2.2.2
Flask-Cors
gunicorn
Werkzeug==2.2.2
requests
python-dotenv
flask-cors
mysql-connector-python
pandas
psycopg2
pytest
pytest-cov
requests-mock
responses
